use global router

+++ src/unity_initiator/cloud/lambda_handler.py
@@ -4,6 +4,8 @@ from tempfile import mkstemp
 from ..router import Router
 from ..utils.logger import logger
 
+ROUTER = None
+
 
 def lambda_handler(event, context):
     logger.info("context: %s", context)
@@ -11,10 +13,12 @@ def lambda_handler(event, context):
     # TODO: Should use either AppConfig or retrieve router config from S3 location.
     # For now, reading router config body from ROUTER_CFG env variable then writing
     # to local file.
-    router_cfg = os.environ["ROUTER_CFG"]
-    fd, router_file = mkstemp(prefix="router_", suffix=".yaml", text=True)
-    with os.fdopen(fd, "w") as f:
-        f.write(router_cfg)
-    router = Router(router_file)
-    os.unlink(router_file)
-    return router.execute_actions(event["payload"])
+    global ROUTER
+    if ROUTER is None:
+        router_cfg = os.environ["ROUTER_CFG"]
+        fd, router_file = mkstemp(prefix="router_", suffix=".yaml", text=True)
+        with os.fdopen(fd, "w") as f:
+            f.write(router_cfg)
+        ROUTER = Router(router_file)
+        os.unlink(router_file)
+    return ROUTER.execute_actions(event["payload"])
